Fix valgrind error seen with saving pdfs with garbage collection.
This was exposed by the recent commit to improve garbage collection.

When we call renumberobjs, previously this might have shortened the
lists without updating the record of how long the lists are.

Now we use expand_lists so that all the lists stay the same length,
never get shorter, and the record is always updated. This also keeps
the +3 nastiness in just one place.