Table hunt: Rework handling of spaces in content.
Previously, we handled spaces in text block by a slightly
arbitrary set of rules when deciding where content was/wasn't.

Single spaces were 'ignored' for gap detection, because we'd
rather see sentences as long unbroken regions of content than
as regions broken at every space. Double (or more) spaces were
accepted as genuine breaks.

Sadly, this falls down with examples such as:

 P subtelomeric                 24.38 8.15  11.31 0.94   1.46
                                11.08 6.93  15.34 0.00   0.73
 Pericentromeric region; C band 20.42 9.26  13.64 0.81   0.81
                                16.50 7.03  14.45 0.17   5.15

Here, the gaps between the first 2 columns of numbers are given
as single spaces, and even the gap between the text of line 3
and the columns of figure is a single space.

We therefore revise the mechanism slightly.

We walk the text, looking for the X coords where text starts and
stops, and store these in a list of tuples:

  (position, frequency, left or right)

We now expand those tuples to include a 'weak' flag too.

The 'weak' flag is used to indicate that all the edges at
this position came from spaces. (So a 'strong' edge and a
'weak' edge at the same place clear the 'weak' flag).

We take the opportunity to improve the handling of text that
either moves left to right, or spills onto multiple lines.

When looking at content to see whether content spills across
between cells (causing them to be merged), we still consider
single spaces as part of the sentence - unless the surrounding
text is numeric. We may revisit this logic at some point.