Table hunt: Be more careful not to leave orphaned bits of stext.
When we move content from one part of the stext tree to another,
avoid leaving empty lines/blocks etc.

Also, when we decide to merge a table into a higher level one
(typically because 2 different parts of the same table were
split into 2 different segments of the page), ensure we don't
leave stray vestiges of the old table around.