Allow pdf_lookup_page_number_slow on deleted pages.
If a page cannot be found in the page tree, return -1 instead of
throwing an error.

This is needed if we disable the fast case for page number lookups
when syncing open pages (where we need to recompute the page number
for currently held pdf_page objects, even those that have been deleted
and removed from the page tree).