Table hunt: Only look for tables inside DIV or at the top level.
Previously, we were looking inside any structure member, such
as 'H', which meant we could oversegment, and then refuse to
find real tables at a higher level.

Also, don't go hunting for tables in a block that  looks  like
a simple column of text. This avoids problems  where  we  have
lines    with    gaps    due    to    extreme    justification
(like that!) that might look confusingly  like  columns  in  a
table.