Paragraph breaking: Allow for whitespace.
SText blocks have a bounding box in them, but for text blocks
this includes any gaps for whitespace.

When figuring out things like "is this paragraph justified",
we can't just look at whether the text line bboxes match the
block bbox edges, as both lines and blocks might have whitespace
around the edges.

For instance, in PMC2656817_00002.pdf we have a paragraph whose
lines end:

  withdrawal from "
  e expression of "
  ure of this im-"
  ed, substantial "

While that is a fully justified paragraph, we failed to detect
it as such, because the whitespace confused us.

In this commit, we calculate the actual 'marked' area (i.e.
ignoring spaces), and modify the logic to use that as appropriate.