Table hunt: Tweak handling of headers/justified text
When looking for the bounds of a table, we were previously
ignoring any 'DIV' that was entirely justified text blocks
or headers.

This was falling foul of PMC1241820_00001.pdf, where we have
a large DIV that contains both 2 text columns and a table
that spans the 2 columns. (The thinness of the gap between
the table and the bottom of the columns was causing us not to
segment across there).

Accordingly, we were proceeding to try to find a table that
spanned all that content.

The fix here, is to ignore any fully justified text or any
headers immediately followed by fully justified text when
looking for table bounds.