jni/android: Ensure that variable fonts variants are also found.
E.g. Android 10 and older had a fixed font for Armenian, but
since Android 11 only provides a variable font for that script.
The original font was named NotoSansArmenian-Regular.otf while
the variable font is named NotoSansArmenian-VF.ttf

So MuPDF must search not just for NotoSans<stem>-Regular, but
also NotSans<stem>-VF when looking for fonts to ensure that it
finds all the fonts in every Android version.