Table hunt: Improve cell vector content detection.
Tidy the code slightly by inlining 'check_for_grid_lines' within
do_table_hunt. That code was not just checking for grid lines.
Logic flow is simpler now.

Rename walk_blocks to be walk_to_find_content, as that's a clearer
name.

Rename erase_grid_lines to calculate_spanned_content. This now
returns the a count of 'duff' cell contents. This includes both
vectors that don't plausibly look like cell contents, and text
that overlaps the edges of the table (Credit to Harald for this
idea!).

The upshot of this commit is that 000000-tables.pdf now correctly
picks up rows/columns that just contain vector graphics as cell
contents.