Bug 708266: Improve font Ascent/Descent handling.
In a recent commit (d0b843e675a6e1f77207f525413f8a0e14bcc4b1),
we changed to use the Ascent/Descent values from the font
descriptor in preference to those from the font itself.

This has produced some issues, because different PDF producers
sometimes put this value as positive, despite it clearly
being specified in the PDF 1.7 spec that it should be negative.
This has in fact been clarified in the 2.0 spec.

Accordingly, if we use this value, and we find a positive
one, we negate it upon use.

Also, we were looking at fontdesc->ascent or descent == 0 to
mean "not present" (it's a compulsory field, it really should
be present!). So rejig to allow for this.