Paragraph breaking: Fix justified text detection with tabs.
In 4-tables.pdf, we have a text fz_stext_block that contains
several fz_stext_lines on the same horizontal line.

These lines have a large gap between them, because they
represent tabs between columns of a table.

Other files have justified blocks in which have lines that
have large gaps between words. This is because large  gaps
are   required    to    justify    that    line    because
(disestablishmentarinism) of the size of long words.

In an attempt to cope with this, we revise the way we spot
justified paragraphs. As we walk each line we gather the
size of the spaces on that line; as long as they are about
all the same, we're happy.

We allow for double spaces as some layout engines put
double spaces after full stops.  Like this.

Finally, the only other tricky bit here is in how we
find space sizes. We have to allow for mixed l2r and
r2l directionality of stuff. For instance, if the
logical text "abc def ghi" is output and laid out to
be "abc fed ghi", then we can't just take (for instance)
the gap between c and d as being a space.

We have to collect fragments and then figure out the
gaps (if any) between fragments.