Paragraph breaking: Tweak list detection.
Firstly, treat a lone 0xFFFD as a bullet point; this is seen
with wingdings symbols.

Secondly, fix detection of ':' and ')' after bullets. Take
care not to recurse too often through is_bullet; '(a)' and
'a)' are reasonable bullets. '((a))' is not.

Thirdly, common up the handling when we find a space after
a potential bullet. (The existing cases in the code had
deviated here).

Finally, spot the case when the entirety of a block is a
bullet. We'll need higher level handling to cope with this
properly.