Paragraph breaking: Improve top level loop.
Firstly, we add some optional helpful debugging so we can easily
see how far we get processing each line.

Next, some of our routines change text blocks into struct blocks
(for instance when we spot that blocks are headers). Also, some
routines split a block into two.

In both such cases, we want to continue to process these blocks -
but not to reprocess them as headers again. Add an 'in_header'
argument to the recursion, and change the next block handling
to allow for this.