Table hunt: Merge 'straddling lines' into line above.
This is to cope with common cases like:

 +-----------------+----------+--------+
 |                 |          | THREE  |
 |                 | TWO LINE |        |
 | ONE LINE HEADER |          | LINE   |
 |                 |  HEADER  |        |
 |                 |          | HEADER |
 +-----------------+----------+--------+

(imagine those lines squashed together a bit so they overlap -
up against the limitations of ascii art here!)

This means we'll detect the table as:

 +---+---+---+   (where v denotes crossing of cell boundary)
 |   |   | # |
 +   +   + v +
 |   | # | # |
 +   + v +   +
 | # | # | # |
 + v +   + v +
 | # | # | # |
 +   + v +   +
 |   | # | # |
 +   +   + v +
 |   |   | # |
 +---+---+---+

A reasonable extraction of this (which makes semantic sense)
would be to replace this with:

 +---+---+---+
 | # | # | # |
 +   +   +   +
 |   | # | # |
 +   +   +   +
 |   |   | # |
 +---+---+---+

or maybe:

 +---+---+---+
 |   | # | # |
 +   +   +   +
 | # | # | # |
 +   +   +   +
 |   |   | # |
 +---+---+---+

We attempt to cope with this by spotting 'straddling rows',
defined to be rows for which every cell is either empty, or
is 'crossed into' from above or below. We are happy to merge
such rows with their neighbours.