jni: Tidy up Archive/TreeArchive/MultiArchive interfaces.
Previously it was impossible to create an empty TreeArchive()
because there was no such constructor, and the underlying struct
created by Archive() was a fz_multi_archive while the Java class
was a plain Archive.

This has now been fixed at the expense of calls to the Archive()
constructor that now needs to go to the MultiArchive()
constructor.