Bug 708286: Improve PDF repair code.
We fix 2 issues here.

Firstly, the selection of a root object. During a repair, we scan
through the file looking for XRef objects that refer to 'Root'.
We pick the 'best' one of these by looking for the last one in
the file, as long as it's really pointing to a dict object.

Sadly, we are doing this before unpacking the ObjStm objects, so
if the last one in the file is in an ObjStm, we'll ignore it in
favour of a previous one.

We fix that by storing the roots we find into a list, and delaying
using that list until after the ObjStm scan.

Secondly, when we process the ObjStms, we replace every object
that exists in the xref with the objects we find in the ObjStm.

This fails to do the right thing if the ObjStm version of the
object comes before non-ObjStm version (or if the ordering of the
objects is unexpected).

We add a simple offset check to ensure we always keep the latest
version.

Finally, we take the opportunity to shuffle functions between files
and refactor slightly so that the number of exposed internal
functions is reduced.