Free unopened pages instead of waiting for document to reap them.
fz_drop_page() decrements a page's reference count, if it reaches
zero, the page is marked dead (by setting its document pointer to
NULL) and left to be freed when the document is freed.

fz_drop_document() iterates over its list of opened pages, and
any pages marked dead (by their document pointer being NULL) are
freed.

fz_load_chapter_page() calls a document type-specific callback to
load a page, before adding the loaded page to the document's list
of opened pages.

The document type-specific callback for PDF, pdf_load_page_imp(),
allocates a pdf_page and then proceeds to load the page's
annotations. This page will only be added to the document's list
of opened pages when pdf_load_page_imp() has finished.

If an exception is thrown while processing the page's
annotations, the page object is dropped.

In this case fz_drop_page() decrements the page's reference count
to zero, assuming that the page will be dropped when its document
is freed, the page is marked dead as usual. However, since the
page was never added to any document, the page will never be
freed, leading to a memory leak.

The fix is for fz_drop_page() to be able to identify pages that
have never been added to their document's list of opened pages,
and it those case, instead of deferring to fz_drop_document() to
reap the pages, just free the pages immediately.

(Problem and solution found by Sebastian!)