Fix assert failure when cleaning a file with object streams.
As seen with:

  mutool clean -dif bug708289.pdf out.pdf

If a file has object stream objects in it, and we try to write
this out non-incrementally, currently, those objects remain as
type 'o'. This leads to an assert failure in the writing code.

A simple fix would be just to write such objects out as 'n'
objects. This would work for non-incremental, non-objstm writes,
at the cost of leaving the ObjStm objects (now unused) in the
file.

A better solution is to do a pass through caching all those
'o' objects in memory, then do another pass, marking them as
'n' objects, and marking the ObjStm objects that they were
loaded from as free.

Includes fixes from Sebastian. Thanks!