PDF saving: Perform a pre-pass to load objects before saving.
With some files, we can trigger a repair during the writing out
of objects (seen with the test file from bug 708289).

We'd rather the repair happened BEFORE we started writing objects.
So perform a pre-pass at the start of the save operation to cache
objects. We would have cached all these objects at some point during
the save operation, so it's not really costing us anything extra.