Improve stext detection of rectangular paths.
When collecting vectors in the stext device we have previously
detected that paths are rectangles and flagged this. This helps
users such as the table hunt code sensibly distinguish
cell-filling background graphics (or borders) from other vector
content.

This works well, when we get content from the PDF like:

   x0 y0 m x0 y1 l x1 y1 l x1 y0 l h f
   x2 y2 m x2 y3 l x3 y3 l x3 y2 l h f
   x4 y4 m x4 y5 l x5 y5 l x5 y4 l h f
   x6 y5 m x6 y7 l x7 y7 l x7 y6 l h f
   ...

as the device sees a series of paths, each of which is a
rectangle.

This falls down, however, when we get data in the PDF like
this:

   x0 y0 m x0 y1 l x1 y1 l x1 y0 l h
   x2 y2 m x2 y3 l x3 y3 l x3 y2 l h
   x4 y4 m x4 y5 l x5 y5 l x5 y4 l h
   x6 y5 m x6 y7 l x7 y7 l x7 y6 l h
   ...
   f

or even:

   x0 y0 m x0 y1 l x1 y1 l x1 y0 l
   x2 y2 m x2 y3 l x3 y3 l x3 y2 l
   x4 y4 m x4 y5 l x5 y5 l x5 y4 l
   x6 y5 m x6 y7 l x7 y7 l x7 y6 l
   ...
   f

Here we get a single path that contains a series of rectangles.
The current code does not detect that as a rectangle, even when
the final area is rectangular.

The fix for this, implemented here, is for the stext device to
walk vectors that it is given, looking for rectangular segments.
If it finds them, these are pulled out as separate stext vector
entries.