Table hunt: Reworking table line detection
Previously, we'd used grid lines to reinforce and refine existing
splits. If a grid line fell outside the max/min regions for a
divider we'd ignore it.

This lead to problems where headers that spanned columns with
spaces in, could bias the position of the divider badly.

For example:

 +-------------+
 | Packaged in |
 +------+------+
 |   A  |   B  |
 ...

With this, content analysis would put the expected split between
the A and B columns aligned with the space in 'Packaged in'.

This would mean that any background rectangles or stroked edges
aligned with the true central point would be confusing to the
algorithm.

Here, we simplify the code that walks the grid lines significantly
so that if we see a rectangle that doesn't align nicely with one
of the existing division regions, we add a new one.

This will result in the above table being seen as being 3 columns
rather than 2, but our merging takes care of this.

Another case that causes us confusion is where we have table cells
like:

 +---------+
 |    A    |
 |         |
 +---------+

Typically, producers will render the background cell with one color
and then may do another background fill just for the character,
resulting in regions like:

 +---+-+---+
 |   |A|   |
 |   +-+   |
 +---------+

To cope with this we amend our merging code to be happy to merge
completely empty columns with neighbours.

This leads some tables with empty tickboxes in to be detected as
tables with empty cells, e.g.

 +---------+--------+
 |         |        |
 +         +  +--+  +
 | Class A |  |  |  |
 +         +  +--+  +
 |         |        |
 +---------+--------+

We now spot such 'bordered empty cells' as actually being
unbordered full ones, and allow our merging routines to shrink
the excess away.

If we pick up a rectangle as being an empty cell, then
allow us to recognise that as a full unbordered cell.