Ensure that cfb archive entry names are null-terminated.
Previously open_cfb_archive_with_stream() depended on that input data
from the file contained a null-terminator for the UTF-8 decoded string
in its internal archive state to be null-terminated.

The parser should always null-terminate its own strings, and not depend
on input data to be well-formed.

This fixes oss-fuzz issue 391935579.