Bug 708293: Allocate xml root node in pool.
Because of the recent commit that uses flexible array members for
the xml tag name the root node no longer contains a 1 byte '\0' to
indicate the empty string for the root node tag name.

Yet, if close_tag() is called after the last tag has been closed,
then the root tag name will be accessed, causing an ASAN complaint.

A simple fix is to allocate the root node of the correct size including
a trailing '\0' for the tag name. Instead of allocating this on the
stack, we now allocate it in the parser pool, which means it will be
dropped whenever the parser and its pool is dropped.