jlib: fix setting filename log level
Building using JLIB_log_levels for filenames only currently fails.

For example, building with the following command:
make JLIB_log_levels=cpp.py=-5 python

... leads to the following trace:

  File "/home/rme/Documents/mupdf/scripts/jlib.py", line 427, in log
    level += log_levels_find( caller)
             ^^^^^^^^^^^^^^^^^^^^^^^^
  File "/home/rme/Documents/mupdf/scripts/jlib.py", line 350, in log_levels_find
    if item_filename and not filename.startswith( item_filename):

This is because ffl is always a list (unpacked in case len==2), and
the filename is the first element of this list.
To fix it, only keep the first element of ffl when there is exactly
one.

Signed-off-by: Raphaël Mélotte <raphael.melotte@mind.be>