Properly stub barcode API when barcode encoding/decoding disabled.
fz_barcode_type_from_string() and fz_string_from_barcode_type() are
missing from the API when barcode support is disabled, which causes
the bindings not to build properly since they are using the missing
fz_barcode_type_from_string() interface, which in turn causes the
bindings documentation to fail to build.

Additionally fz_barcode_type_strings is declared but never used when
barcode support is disabled, which causes a warning.