Bug 708304: Handle PDF objects numbered outside xref range.
The PDF audit tool would trigger an ASAN warning when encountering a
PDF object outside of the xref range, since the audit tool allocates
internal state based of the number of objects specified by the xref.
Later on when encountering an object that resides outside of that
range, the audit tool did no bounds checking.

The fix is to add bounds checking to avoid accessing outside the
internal state.