Improve how we provide custom images to stamp annotations.
Detect the case when a Stamp annotation appearance has a resource
dictionary with a single Image defined. Use this detection to provide
functions to get and set a custom image for stamp annotations.

This allows us to delay creating the appearance stream, and will allow
us to change other appearance properties such as the opacity (CA) without
clobbering the image.