MSVC: Up the warning level in most projects.
This is to try and get a similar level of warnings to that given
in unix builds. In particular "defined but not used variables"
and "set but not used variables" are reported on Linux, but not
on MSVC, which often leads to fixup commits.

The approach taken here is to up the warning level from 3 to 4
for all our code, but not for the thirdparty stuff.

This produces many more warnings than we'd like - indeed lots
of warnings that we don't actually want. So we then specifically
hide the annoying ones on a per project basis.

In addition, we hide specific warnings so that the thirdparty
stuff doesn't give us more warnings than we want.

This still leaves us with lots of warnings more than we had before.
Expect a series of commits to gradually squash these away,
leaving the code in better shape than before.