Update fz_format_string for XML and hexstrings.
Fix the implementation of "%<" (format a string for XML output)
to properly cope with XMLs limitations in terms of the characters
it can represent.

Add some helpful functions to detect chars that are out of range
or strings that contain chars that are out of range, etc.

Add a new "%>" specificer to format a string into a hexstring.
This is valid in all cases and can serve as a helpful fallback
when users detect that the XML encoding is "lossy".