Bug 708323: Workaround PDFs with bad Structure Trees in.
The example in the bug is from Microsoft Powerpoint 365, but
whether this has been manipulated since being produced is
unclear.

If we are processing a page, and we come across a MCID 23 (say),
then we are supposed to lookup the StructParent from the page
object (a number). We lookup that number in the ParentTree number
tree within the StructTreeRoot. This should point to an array
of objects.

The idea is that index 23 of this array should be the StructElem
we want.

BUT... in the example file, this isn't. Some MCIDs have no
matching element in the array. This means that index 23 of the
array does not give the one we want.

So, the fix here is to lookup the object in the way we'd expect,
and then to double check that that is the object required (by
looking in the /K entry). If it is not, we exhaustively search
the array until we find one that is correct.