Add FZ_STEXT_ACCURATE_SIDE_BEARINGS options.
If set, this ensures that the char bboxes returned in the stext
device completely cover the width of the glyph outlines, rather
than just covering from the origin to the advance.

Note that the width returned will always cover the origin to
advance position, even if the outlines are smaller.

Couple this with FZ_STEST_ACCURATE_ASCENDERS and you are guaranteed
to get a bbox that completely covers the glyph outlines.