Bug 708393: Fix pdf_count_q_balance.
pdf_count_q_balance relies on counting q and Q operations
to detect mismatches.

Unfortunately, it does this by running the standard PDF
interpretation mechanisms, and these attempt to autocorrect
for missing Q's at the end of a page. This autocorrection
affects our reading.

Insert a new processor event that happens before the auto
correction, and use that so that pdf_count_q_balance can get
the uncorrected values.