Bug 708442: Avoid crash due to unallocated stroke state.
Commit 5576bba26c0fe38243b25365c6b0faf0638114b5 accidentally omitted
allocating a new stroke state when a new line is allocated. ASAN
detects the NULL dereference that causes a crash.

The fix is to ensure that the stoke state is allocated when a new
line is allocated. This is reasonable since the original code just
depended on fz_default_stroke_state. After this fix ASAN no longer
complains.