Take note of Intent field in PDF Images.
This is supposed to override the intent in the graphics state.

Update fz_image objects to note whether we have an intent, and
if so, what it is.

Update the calls from the PDF interepreter into the rendering
code to use these values. This means that we retain the ability
in the device interface to display an image with a different
intent to that which is specified, as this seems more general.