Bug 708395: Fix a couple of issues with PCL output.
Firstly, when skipping blank lines, we were forgetting to
increment the data file. This means that any colour output file
that starts with a black link would just be printed empty. This
fixes input1.pdf.

Secondly, tweak the page margin setup code. I think we'd dropped
a character from this setup at some point. For clarity, split
the different commands to be separate and add comments.