Bug 708489: Fix bugs in TTF subsetting.
When subsetting the POST table, we were omitting entries for
glyphs that had to be included because they were used as
composites. For example, in the given file, the ':' glyph
uses the '.' glyph twice. The '.' glyph would otherwise be
removed.

I'm not sure if this actually causes a problem, but it's nicer
to include it.

The second issue, and the one that actually causes the problem
is the regeneration of the cmap table. We were incorrectly
checking the char id against the number of glyphs, rather than
the maximum of the encoding.

We were getting away with this as in most cases the number of
glyphs is still quite large. In this example, however, we were
dealing with a font that had already been subset. So the char
id's were regularly larger than the number of glyphs in our
"original" font.