Tweak detection of epub files.
Previously we detected the existence of META-INF/container.xml
as being a 100% indication that we had an epub file.

This is bad because HWPX files also have such files, and they
contain content that make us much more sure that they are
HWPX files - and we only score those as 75% to enable
sodochandler to score them higher if it is enabled.

Accordingly, reduce epubs confidence to 74%.