ttf: Iterate over old instead of new strings when subsetting post subtable in TTF.
The last loop in subset_post2() intended to iterate over all prior
strings and compare their stringindices with heap.a to see if each
string index exists in the heap and therefore should be kept.

Unfortunatley the loop limit, n, was set to the length of the heap.
So for a file where the number of strings was e.g. 230, but where
the number of strings to keep was 69, the loop would terminate after
69 entries without even checking the latter strings (likely reducing
the number of strings to less than 69). The string count stored in
the post2 header would still be 69, and the glyph name index would
still keep space for 69 entries. This leads to out of bounds accesses
causing rendering problems with PDFs that use subset TTF with post2
subtables.