Bug 708509: Return index into ocg list after binary search.
Previously find_ocg() returned the difference between the OCG
object number being searched for and the one found. This resulted
in populate_ui() always setting ui->ocg == 0 (when the ocg was
present), which finally meant that pdf_layer_config_ui_info()
always set every OCG's selection state to that of the first OCG.

The fix is to simply return the index into the array after the
binary search finds an entry.