Bug 708515: Revamp recognizer behaviour slightly.
Currently, when asked to recognize a stream (or file), we find the
recogniser that gives the highest score for recognising the file
from its content.

If this score is less than 100, we then look at the 'magic' value
supplied. If this matches, we then override the choice with that.

This means that if you say "open an html file", with magic="jpg",
the html recogniser scores it as 75 (leaving room for a hypothetical
better match to score higher). The magic='jpg' then causes it to
actually be opened using the image handler, and it will fail.

Instead, in this commit, we move to a scheme where magic is used
to a) choose between format recognizers that give a non-zero score,
or b) to pick a format when no recognizer gives a non-zero score.

So if the HTML and XHTML recognisers both recognise a file, magic
will be used to choose which one is used. But magic can never cause
a JPG file to be loaded as an HTML one.