Fix handling of offsets in fz_extract_html_from_mobi.
We run through the file reading offset records. These are assumed
to be monotonically increasing, so we intend to move minoffset
up each time.

Unfortunately the logic to do this is broken, both because 'min'
is the wrong thing to use, and because i can be larger than k and
so into uninitialised data.

Use a simpler updating technique as discussed with Sebastian.