Bug 708545: Fix strikeout/underline detection.
The existing code, upon encountering a rectangle, would look
through the text to see if the rectangle struckout or
underlined some text.

This fails with the given file because the underlines/strikeouts
are sent BEFORE the text.

So, the fix is to store the lines in a list, and then to check
them all at device closedown.