Modify stext structure to represent multiple pages.
For the purposes of extracting structured output, we already allow
multiple extractions to take place into a single fz_stext_page.
The structured information collects properly over complete documents,
but we lose knowledge of which bits came from which page.

To try to solve this, we add an 'id' value to every fz_stext_block.
This id value should be an index into a new array of
fz_stext_page_details held in the top level fz_stext_page structure.

These details include the chapter and page number from which
the page was extracted, together with the mediabox for that page.

Thus, for every bit of data in the stext, we can figure out what
page it came from.

In theory this should allow us to search the complete document
looking for results that span pages.