Add fz_match... functions.
These do the same as fz_search functions, but rather than just
doing 'case-insensitive' matching, they allow for a wider range
of text transformations, and regexp searches.

At the simplest level, people can ask for matching to be performed
'exactly', 'case-insensitively', or 'forgiving'.

Forgiving searches are performed by doing a unicode compatible
decomposition, and then filtering out "marking, non-spacing"
characters. The hope is that this should allow A-acute to match
A, etc.

People wanting to control this more closely can use the inner
fz_match_stext_page_cb_with_finder function to control the
exact unicode decompostion/transform/recomposition processes
that take place.

Parts of this were stolen from a work-in-progress by Tor, and
credit is due to Sebastian for the decomposition idea.

Extend searching with different whitespace flattening options.

When the haystack is differently flattened, we need to adjust our
routine that walks the stext "in step" with the haystack.

Update searching for chapter/page.