Fix mistake in reworked recognition code.
If we have 2 DAs that claim to support the same "magic" (such
as mupdf's core docx handler, and the docx handler that relies
on SmartOffice), we want the 'better' one to be used.

The recent tweaks to recognition broke this. Fix it here.

Also, don't stop looking for magic_score, just because we have
a non-zero score.