Revise subsampling code.
If we subsample an image that's not a multiple of the subsample
factor in size, we end up sampling the last row/column multiple
times. This has the effect of 'shifting' the image in one direction.

Here, we tweak the subsampling so these repeated row/columns are
spread more evenly through the image.