Strikeout detection: Extend rectangles to improve detection.
Spaces and underscores with accurate-text can leave zero (or
small) height rectangles, which makes strikeout detection poor.
We already fudge the rectangles down for underline detection.
Extend this to fudging it upwards too for strikeout - but not
so far that we end up including cell borders.

Also, ensure we check the thickness of rects; if they are more
than a quarter of the height of the font, assume they are a
highlight rather than a strikeout!