Don't decode full JPX image when loading metadata only.
Look at opj_image.color_space to determine number of components/alpha.

The opj_comp.alpha field is only filled in AFTER decoding an image.

If we want to load JPX metadata without decoding the full image, we cannot rely
on those fields to be filled out. Look at the color_space field, default
colorspace (from PDF) and compare with numcomps to detect the number of
components and alpha channels instead.

This sometimes fails when the color_space is unknown/unspecified, because
we assume RGB instead of CMYK to match how Adobe handles these situations.

This causes a mix of regressions and progressions in some
tests_private/j2k/openjpeg/input/conformance files; some CMYK files are treated
as RGBA, and vice versa.