Followup to recent subsampling commit.
If we subsample a large enough amount, we can access off the end
of the image. Seen with:

 mutool draw -o out.png ../PyMuPDF.git/tests/resources/test_3727.pdf

under valgrind.

There we try to subsample a 1000x1000 image by a factor of 1<<6 = 64.

We want to treat that as a 1024x1024 image, which means we need to
step back 24 times as we walk across the block; but we only have
16 blocks. Thus we sometimes need to step back twice; hence use
"while", not "if" in the bresenham step.

We also want to step 24 times by the time we reach the last block - so
in 15 steps, not in 16. Therefore we should be setting up the
bresenham to step in 1 block fewer.