Fix strikeout detection failure caused by FP inaccuracy.
When we try to figure out if a line intersects a rectangle,
we do it by 'advancing' the start of the line to the edge
of the rectangle. This advancement was, in some cases, not
actually quite reaching, due to FP inaccuracies.

Reformulate it slightly so we just assign the appropriate
FP value rather than perform operations on it.

This was showing up with sections of PyMuPDFs test_4503.pdf
missing the strikeout detection.