Bug 708587: Be more accepting of EmbeddedFiles FileSpecs.
Previously we used to insist on matching the scheme used for the
name and the filename for embedded files. (So if we find the
name as /UF, then we'll look for /EF << /UF ... >> rather than
anything else.

The example file in this bug shows that some (bad!) producers are
failing to match the two.

So, in the event we can't find a matching pair, accept a
non-matching pair.