Bug 708594: Improve speed of ps_roll.
This is the heaviest obvious hotspot in the rendering, but it's
not that heavy, so it's not a huge improvement.

BUT... we were memcpying n things rather than n-1 each time around
the loop (and that's bad), so we might as well fix that. And while
we're in there, let's make the common case of rolling upwards
rather than downwards fast too.